feat: style props without box
