+++ docs/components/example/bottom-sheet-preview.tsx
@@ -16,8 +16,7 @@ const BottomSheetPreview = () => {
         <ActionButton>Open</ActionButton>
       </BottomSheetTrigger>
       <BottomSheetContent title="제목" description="설명을 작성할 수 있어요">
-        {/* BottomSheetBody is simple paddingX helper. */}
-        {/* If you need to omit padding, replace with <Stack /> or other layout you need. */}
+        {/* If you need to omit padding, pass paddingX={0}. */}
         <BottomSheetBody minHeight="x16">Content</BottomSheetBody>
         <BottomSheetFooter>
           <ActionButton variant="neutralSolid">확인</ActionButton>
