+++ packages/qvism-preset/src/recipes/bottom-sheet.ts
@@ -100,7 +100,18 @@ const bottomSheet = defineRecipe({
       display: "flex",
       flexDirection: "column",
 
-      paddingInline: vars.base.enabled.body.paddingX,
+      "--seed-box-padding-x": vars.base.enabled.body.paddingX,
+      "--seed-box-height": "initial",
+      "--seed-box-min-height": "initial",
+      "--seed-box-max-height": "initial",
+      "--seed-box-justify-content": "initial",
+      "--seed-box-align-items": "initial",
+      paddingInline: "var(--seed-box-padding-x)",
+      height: "var(--seed-box-height)",
+      minHeight: "var(--seed-box-min-height)",
+      maxHeight: "var(--seed-box-max-height)",
+      justifyContent: "var(--seed-box-justify-content)",
+      alignItems: "var(--seed-box-align-items)",
     },
     footer: {
       display: "flex",
